Issue/forum threads sorting (#317)
* made call to PageActionsButtons more conform to convention

* fixed layout of FilterBar in forum/categories/category/threads

* buildkite started complaining about yarn.lock

* issue seems to be resolved.
Took a peak at how routes/forum/categories/category/threads/thread/posts/index handles the loading of the model, since every post belongs to one thread, similar to how every thread belongs to one category. Applying the same technique for querying based on a filter in params was pretty straightforward, but required the addition of a filter in the backend. See the equally named branch in amber-api for this addition.
Note: the params attribute in the threads/index route is somewhat different from the params in the posts route; the corresponding posts controller does not extend filterableAndSearchable, whereas the threads controller does extend that mixin.

* fixed indent of tags to make test pass

* linter

Co-authored-by: matteo <m.bronkhorst-1@student.utwente.nl>
Co-authored-by: Wilco <wilcovanbeijnum@gmail.com>
Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>